Templates model updated

+++ zeus/modules/deployer/src/main/resources/zeus-deployer/utils/application/Variables.js
@@ -3,8 +3,8 @@ var dao = require('zeus-applications/data/dao/Variables')
 exports.create = function(applicationId, deployment) {
 	var containers = deployment.spec.template.spec.containers;
 	for (var i = 0; i < containers.length; i ++) {
-		var env = containers.env;
-		for (var j = 0; j < env.length; j ++) {
+		var env = containers[i].env;
+		for (var j = 0; env !== undefined && j < env.length; j ++) {
 			dao.create({
 				'Name': env[j].name,
 				'Value': env[j].value,
@@ -19,5 +19,7 @@ exports.delete = function(applicationId) {
 		'Application': applicationId
 	})[0];
 
-	dao.delete(entity.Id);
+	if (entity) {
+		dao.delete(entity.Id);
+	}
 };
